requirements.txt
discord.py
mcrcon
psycopg2
